+++ src/parts/CreateRpcWithId/CreateRpcWithId.ts
@@ -1,13 +1,13 @@
-import * as Assert from '../Assert/Assert.ts'
 import * as ExtensionHostRpcState from '../ExtensionHostRpcState/ExtensionHostRpcState.ts'
 import * as ExtensionHostSubWorkerUrl from '../ExtensionHostSubWorkerUrl/ExtensionHostSubWorkerUrl.ts'
 import * as IpcParent from '../IpcParent/IpcParent.ts'
 import * as IpcParentType from '../IpcParentType/IpcParentType.ts'
 import * as RpcParent from '../RpcParent/RpcParent.ts'
 import * as RpcParentType from '../RpcParentType/RpcParentType.ts'
+import * as RpcState from '../RpcState/RpcState.ts'
 
-export const createRpcWithId = async ({ id, execute }: { id: string; execute: any }) => {
-  Assert.string(id)
+export const createRpcWithId = async (id: string) => {
+  const fn = RpcState.acquire(id)
   const info = ExtensionHostRpcState.get(id)
   if (!info) {
     throw new Error(`rpc with id ${id} not found`)
@@ -17,11 +17,12 @@ export const createRpcWithId = async ({ id, execute }: { id: string; execute: an
     url: ExtensionHostSubWorkerUrl.extensionHostSubWorkerUrl,
     name: info.name,
   })
-  const rpc = await RpcParent.create({
+  const newRpc = await RpcParent.create({
     ipc,
     method: RpcParentType.JsonRpc,
-    execute,
+    execute: fn,
   })
-  await rpc.invoke('LoadFile.loadFile', info.url)
-  return rpc
+  await newRpc.invoke('LoadFile.loadFile', info.url)
+  RpcState.set(id, newRpc)
+  return newRpc
 }
