+++ src/parts/ExtensionHostRpc/ExtensionHostRpc.ts
@@ -7,6 +7,7 @@ import * as IpcParentType from '../IpcParentType/IpcParentType.ts'
 import * as RpcParent from '../RpcParent/RpcParent.ts'
 import * as RpcParentType from '../RpcParentType/RpcParentType.ts'
 import { VError } from '../VError/VError.ts'
+import * as GetOrCreateRpcWithId from '../GetOrCreateRpcWithId/GetOrCreateRpcWithId.ts'
 
 const defaultExecute = () => {
   throw new Error('not implemented')
@@ -16,8 +17,7 @@ export const createRpc = async ({ id, url, name, execute = defaultExecute, conte
   try {
     if (id) {
       Assert.string(id)
-      const rpc = await CreateRpcWithId.createRpcWithId({ id, execute })
-      return rpc
+      return GetOrCreateRpcWithId.createRpcWithId({ id, execute })
     }
 
     // deprecated
